Make dependency locking failures lenient
Locking failures no longer cause an exception to be thrown, they are
instead recorded in the resolution result.
This enables lenient consumption of the resolution result to succeed
while preserving the failure behaviour for non lenient use cases.
This is done by recording artificial edges from the root component to
the different locking violations, either a missing module or an
unexpected one.

Issue #6383