Omit excludes for Maven optional dependencies
For BOM import support, when a <dependencyManagement> entry has a declared <exclusion>,
we apply this exclusion globally whenever the target module is resolved, regardless of
the path to the module. However, this behaviour is not desired for <optional> dependencies,
and these could not be distinguished from <dependencyManagement> entries after the POM was parsed.

With this change, we retain a 'type' attribute for all dependencies parsed from a Maven POM
file. This allows us to omit any exclusions that are defined for a dependency declared with
<optional>, while retaining them for <dependencyManagement> entries.